Added Unit Controllers
Implemented controllers for Ball and Brick. Created 'noInput' input class.
