+++ Assets/Scripts/Controllers/BrickController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 132a59b5147c36b4b96d3ccc621d6f0b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
