+++ Assets/Scripts/Controllers/NoInput.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aa066c46531b5ff44b712e63abc71b49
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
